Optimize for deadline handling
Refactored TCPSocket and SSLSocket:
- will not longer overwrite #read and #write
- uses dedicated methods #read_with_deadline and #write_with_deadline (see IOWithDeadlineMixin).

If you are using this socket classes directly then this is a breaking change for you.

This optimizes for a "query in a time frame" use case: it should not matter if the write/send part of a request is in time or if the read/receive part is in time but the complete query/request should be finished within a given time frame.